requirements.txt
Django==2.2
django-sql-explorer==1.1.2
whitenoise==4.1.2
gunicorn==19.9.0
